+++ src/Features/VisualBasic/Portable/Completion/CompletionProviders/CompletionUtilities.vb
@@ -4,7 +4,6 @@
 
 Imports System.Collections.Immutable
 Imports Microsoft.CodeAnalysis.Completion
-Imports Microsoft.CodeAnalysis.Completion.Log
 Imports Microsoft.CodeAnalysis.Completion.Providers
 Imports Microsoft.CodeAnalysis.Options
 Imports Microsoft.CodeAnalysis.Shared.Extensions.ContextQuery
@@ -147,11 +146,6 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.Completion.Providers
                     ' document.
                     Return insertionText.Substring(0, insertionText.Length - 1)
                 End If
-            End If
-
-            If ch = "." AndAlso SymbolCompletionItem.GetShouldProvideParenthesisCompletion(Item)
-                CompletionProvidersLogger.LogCommitUsingDotToAddParenthesis()
-                Return insertionText + "()"
             End If
 
             Return insertionText
